web finish 2.0

+++ app/Http/Controllers/Admin/PostController.php
@@ -84,6 +84,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('author', $post);
 
         $categories = Category::pluck('name', 'id');
 
@@ -101,6 +102,9 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
+
+        $this->authorize('author', $post);
+
         $post->update($request->all());
 
         if ($request->file('file')) {
@@ -135,6 +139,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('author', $post);
+
         $post->delete();
 
         return redirect()->route('admin.posts.index', $post)->with('info', 'Post Delete Successfully');
